requirements (3).txt
streamlit>=1.32.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
matplotlib>=3.8.0
tqdm>=4.66.0
pandas>=2.2.0
scikit-learn>=1.4.0
